Update Made to repo

+++ 2011/gb_mandriva_MDVSA_2011_196.nasl
@@ -24,7 +24,32 @@
 ###############################################################################
 
 include("revisions-lib.inc");
-tag_insight = "A vulnerability has been discovered and corrected in ipmitool:
+
+
+if(description)
+{
+  script_xref(name:"URL", value:"http://lists.mandriva.com/security-announce/2011-12/msg00028.php");
+  script_oid("1.3.6.1.4.1.25623.1.0.831515");
+  script_version("$Revision: 11979 $");
+  script_tag(name:"last_modification", value:"$Date: 2018-10-19 10:21:43 +0200 (Fri, 19 Oct 2018) $");
+  script_tag(name:"creation_date", value:"2011-12-30 09:13:10 +0530 (Fri, 30 Dec 2011)");
+  script_tag(name:"cvss_base", value:"3.6");
+  script_tag(name:"cvss_base_vector", value:"AV:L/AC:L/Au:N/C:N/I:P/A:P");
+  script_xref(name:"MDVSA", value:"2011:196");
+  script_cve_id("CVE-2011-4339");
+  script_name("Mandriva Update for ipmitool MDVSA-2011:196 (ipmitool)");
+
+  script_tag(name:"summary", value:"Check for the Version of ipmitool");
+  script_category(ACT_GATHER_INFO);
+  script_copyright("Copyright (c) 2011 Greenbone Networks GmbH");
+  script_family("Mandrake Local Security Checks");
+  script_dependencies("gather-package-list.nasl");
+  script_mandatory_keys("ssh/login/mandriva_mandrake_linux", "ssh/login/release", re:"ssh/login/release=MNDK_(mes5|2010\.1)");
+  script_tag(name:"affected", value:"ipmitool on Mandriva Linux 2010.1,
+  Mandriva Linux 2010.1/X86_64,
+  Mandriva Enterprise Server 5,
+  Mandriva Enterprise Server 5/X86_64");
+  script_tag(name:"insight", value:"A vulnerability has been discovered and corrected in ipmitool:
 
   ipmievd as used in the ipmitool package uses 0666 permissions for
   its ipmievd.pid PID file, which allows local users to kill arbitrary
@@ -35,52 +60,19 @@ tag_insight = "A vulnerability has been discovered and corrected in ipmitool:
   rather easily craft an initscript or start the service by other means
   rendering the system vulnerable to this issue.
 
-  The updated packages have been patched to correct this issue.";
-tag_solution = "Please Install the Updated Packages.";
-
-tag_affected = "ipmitool on Mandriva Linux 2010.1,
-  Mandriva Linux 2010.1/X86_64,
-  Mandriva Enterprise Server 5,
-  Mandriva Enterprise Server 5/X86_64";
-
-
-if(description)
-{
-  script_xref(name : "URL" , value : "http://lists.mandriva.com/security-announce/2011-12/msg00028.php");
-  script_oid("1.3.6.1.4.1.25623.1.0.831515");
-  script_version("$Revision: 9371 $");
-  script_tag(name:"last_modification", value:"$Date: 2018-04-06 10:55:06 +0200 (Fri, 06 Apr 2018) $");
-  script_tag(name:"creation_date", value:"2011-12-30 09:13:10 +0530 (Fri, 30 Dec 2011)");
-  script_tag(name:"cvss_base", value:"3.6");
-  script_tag(name:"cvss_base_vector", value:"AV:L/AC:L/Au:N/C:N/I:P/A:P");
-  script_xref(name: "MDVSA", value: "2011:196");
-  script_cve_id("CVE-2011-4339");
-  script_name("Mandriva Update for ipmitool MDVSA-2011:196 (ipmitool)");
-
-  script_tag(name:"summary", value:"Check for the Version of ipmitool");
-  script_category(ACT_GATHER_INFO);
-  script_copyright("Copyright (c) 2011 Greenbone Networks GmbH");
-  script_family("Mandrake Local Security Checks");
-  script_dependencies("gather-package-list.nasl");
-  script_mandatory_keys("ssh/login/mandriva_mandrake_linux", "ssh/login/release");
-  script_tag(name : "affected" , value : tag_affected);
-  script_tag(name : "insight" , value : tag_insight);
-  script_tag(name : "solution" , value : tag_solution);
+  The updated packages have been patched to correct this issue.");
+  script_tag(name:"solution", value:"Please Install the Updated Packages.");
   script_tag(name:"qod_type", value:"package");
   script_tag(name:"solution_type", value:"VendorFix");
   exit(0);
 }
 
-
 include("pkg-lib-rpm.inc");
 
-release = get_kb_item("ssh/login/release");
-
+release = rpm_get_ssh_release();
+if(!release) exit(0);
 
 res = "";
-if(release == NULL){
-  exit(0);
-}
 
 if(release == "MNDK_mes5")
 {
@@ -91,7 +83,7 @@ if(release == "MNDK_mes5")
     exit(0);
   }
 
-  if (__pkg_match) exit(99); # Not vulnerable.
+  if (__pkg_match) exit(99);
   exit(0);
 }
 
@@ -105,6 +97,6 @@ if(release == "MNDK_2010.1")
     exit(0);
   }
 
-  if (__pkg_match) exit(99); # Not vulnerable.
+  if (__pkg_match) exit(99);
   exit(0);
 }
